requirements.txt
streamlit==1.32.0
Pillow==10.2.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
mediapipe==0.10.21
google-generativeai==0.8.5
websockets==12.0
aiohttp==3.9.3
python-dotenv==1.0.1 